backend/requirements.txt
openai==0.27.0
Flask==2.1.1
Flask-Cors==3.0.10
PyYAML==5.4.1
